Updated the app to look more pleasant
